writer_svc/requirements.txt
flask==2.3.3
kafka-python==2.0.2
pyyaml==6.0.1
numpy==1.24.3
pandas==2.0.3
lancedb==0.5.0
psycopg2-binary==2.9.7
requests==2.31.0
